No access to any inventories with Introspection module
When a player is restricted, all inventories and transfer locations provided by the Introspection module are seen as completely empty, and pushing items into or pulling from them moves no items, no matter if it is between player inventories or a physical container.

This also fixes Baubles and armor slot support for the Introspection module.

Bumps version to v1.2.0.